requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.4
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.2
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
jinja2==3.1.6
markupsafe==3.0.2
pynacl==1.5.0
pyyaml==6.0.2
requests==2.32.5
urllib3==2.5.0
python-dotenv==1.1.1

